+++ demo/src/Examples/index.js
@@ -1,4 +1,0 @@
-import ExampleMasonry from "./Masonry"
-import ExampleResponsiveMasonry from "./Responsive"
-
-export {ExampleMasonry, ExampleResponsiveMasonry}