+++ example/gatsby-config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  flags: {
-    FAST_DEV: true,
-    DEV_SSR: true,
-  },
-  plugins: [],
-}